[ALPS05574842] wlan: Fix build error
Fix build error.

MTK-Commit-Id: def1ca316f2e7f8b6d575e8454b0d23d927fc2d4

CR-Id: ALPS05574842
Feature: [Module]Wi-Fi Driver
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/5439706
CheckPatch: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_mspautosanity

+++ include/chips/cmm_asic_common.h
@@ -75,7 +75,7 @@ void fillTxDescAppendByHostV2(IN struct ADAPTER *prAdapter,
 	IN struct MSDU_INFO *prMsduInfo, IN uint16_t u4MsduId,
 	IN phys_addr_t rDmaAddr, IN uint32_t u4Idx, IN u_int8_t fgIsLast,
 	OUT uint8_t *pucBuffer);
-void halDumpTxdInfo(IN struct ADAPTER *prAdapter, uint32_t *tmac_info);
+void halDumpTxdInfo(IN struct ADAPTER *prAdapter, uint8_t *tmac_info);
 void asicWakeUpWiFi(IN struct ADAPTER *prAdapter);
 #endif /* _CMM_ASIC_COMMON_H */
